requirements.txt
Django==5.0
Werkzeug==3.0.1
django-cors-headers==4.3.1
django-extensions==3.2.3
djangorestframework==3.14.0
opencv-python==4.8.1.78
python-dotenv==1.0.0
psycopg2-binary==2.9.9